Compilation fixes
git-svn-id: file:///fltk/svn/trunk@5855 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ OpenGL/gl_draw.cxx
@@ -108,7 +108,7 @@ void fltk::glsetfont(fltk::Font* font, float size) {
     short cfont;
     GetFNum(CFStringGetPascalStringPtr(cfname, kCFStringEncodingMacRoman),&cfont);
     CFRelease(cfname);
-    aglUseFont(aglGetCurrentContext(), cfont, attrib,current_size_,0,256,listbase);
+    aglUseFont(aglGetCurrentContext(), cfont, attrib,(int)current_size_,0,256,listbase);
 #else
 #error
 #endif
